Create cy.loadMultiFlatmap()
