+++ package.json
@@ -17,8 +17,8 @@
     "serve": "vite --host --force",
     "build-bundle": "vite build",
     "build-static": "vite build -c vite.static-build.js",
-    "test": "cypress run --component",
-    "cypress:open": "cypress open",
+    "cypress:open": "npm run cypress:prepare && cypress open",
+    "cypress:component": "npm run cypress:prepare && cypress run --component",
     "release:beta": "npm version prerelease --preid=beta; npm publish --tag beta",
     "release:minor": "npm version minor; npm publish",
     "release:patch": "npm version patch;  npm publish",
